Этап 1 ver 1.1

+++ stat-service/stat-server/src/main/java/ru/practicum/statistic/api/storage/StatisticInfo.java
@@ -5,5 +5,5 @@ public interface StatisticInfo {
 
     String getUri();
 
-    int getHits();
+    long getHits();
 }
